tested finalised API endpoints, fixed some bugs, updated return of basket calls to return the basketViewModel having been updated

+++ Checkout/Model/DTO/GetProductViewModel.cs
@@ -6,6 +6,6 @@ namespace CheckoutAPI.Model.DTO
     {
         public virtual long Id { get; set; }
         public virtual string Name { get; set; }
-        public virtual float Price { get; set; }
+        public virtual double Price { get; set; }
     }
 }
